Add support for go modules.
This allows anyone who uses go 1.11 or later to build nri-mysql
integration using the go module support. According to the README.md and
newrelic customer support, contributions from the community are
welcomed. Switching to go modules is the way go code is supposed to
be built and improves developer ergonomics because
I can type go build and it should just work (I don't have to worry about
vendoring, using one of the numerous third party dep tools and so on),
its also faster.

Note because this is a program and not a library there should no
breaking changes. I verified that go build in module mode will
successfully build. To avoid breaking the way make currently
builds the code, I have left the src folder and vendor folder alone.
This change is minimal: it allows go modules to work while not
disrupting the current make workflow.

If there is interested in adjusting the build tooling to make it
a fit a more traditional go workflow for modules (nuke vendor folder,
move code out of src and move them into the root of the directory),
please let me know and I would be willing to make those changes.

See here for details: https://github.com/golang/go/wiki/Modules